+++ app/src/main/java/com/example/borntobe/HandAnalysisActivity.kt
@@ -2,29 +2,51 @@ package com.example.borntobe
 
 import android.graphics.Bitmap
 import android.graphics.Canvas
+import android.graphics.Color
 import android.graphics.ImageDecoder
+import android.graphics.Paint
+import android.graphics.drawable.BitmapDrawable
 import android.os.Bundle
+import android.util.AttributeSet
 import android.util.Log
+import android.view.View
 import android.widget.Button
 import android.widget.ImageView
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
+import androidx.core.graphics.createBitmap
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.example.borntobe.databinding.ActivityHandAnalysisBinding
 import com.google.mediapipe.tasks.vision.core.RunningMode
+import com.google.mediapipe.tasks.vision.handlandmarker.HandLandmarker
+import com.google.mediapipe.tasks.vision.handlandmarker.HandLandmarkerResult
+import kotlin.math.max
+import kotlin.math.min
 
 class HandAnalysisActivity : AppCompatActivity() {
+    companion object {
+        private const val LANDMARK_STROKE_WIDTH = 8F
+    }
+
+    private var results: HandLandmarkerResult? = null
+    private var linePaint = Paint()
+    private var pointPaint = Paint()
+
+    private var scaleFactor: Float = 1f
+    private var imageWidth: Int = 1
+    private var imageHeight: Int = 1
+
     // ViewBinding 사용
     private lateinit var binding: ActivityHandAnalysisBinding
     private lateinit var ivImage: ImageView
     private lateinit var btnImageUpload: Button
     private lateinit var btnAnalysis: Button
-    private lateinit var bitmap: Bitmap
+    private lateinit var image: Bitmap
     private lateinit var handLandmarkerHelper: HandLandmarkerHelper
-    private lateinit var overlayView: OverlayView
     private val utils: Utils = Utils(this)
 
     // isPhotoPickerAvailable : 사진 선택 도구가 사용 가능한 기기 인지 확인
@@ -47,7 +69,8 @@ class HandAnalysisActivity : AppCompatActivity() {
                 utils.showToast("이미지가 업로드 되었습니다.")
                 // uri를 bitmap으로 변환
                 val source = ImageDecoder.createSource(this.contentResolver, uri)
-                bitmap = ImageDecoder.decodeBitmap(source)
+                val bitmap = ImageDecoder.decodeBitmap(source)
+                image = bitmap.copy(Bitmap.Config.ARGB_8888, true)
             } else {
                 Log.i("PhotoPicker", "No media selected")
                 utils.showToast("이미지를 선택하세요.")
@@ -82,25 +105,91 @@ class HandAnalysisActivity : AppCompatActivity() {
         btnAnalysis = binding.activityHandAnalysisBtnAnalysis
         btnAnalysis.setOnClickListener {
             // 손 분석을 위한 HandLandMarkerHelper 객체 생성
-            handLandmarkerHelper = HandLandmarkerHelper(
-                0.5F,
-                0.5F,
-                0.5F,
-                1,
-                0,
-                RunningMode.IMAGE,
-                this)
+            handLandmarkerHelper = HandLandmarkerHelper(context = this)
             // 손 분석 결과로 추출된 LandMark 반환
-            val handLandMarkerResult = handLandmarkerHelper.detectImage(bitmap)
+            val handLandMarkerResult = handLandmarkerHelper.detectImage(image)
+            // 결과값이 null이 아니면 landmark 그려서 보여줌
             if (handLandMarkerResult != null) {
-//                val canvas = Canvas(bitmap)
-//                overlayView = OverlayView(this, null)
-//                overlayView.clear()
-//                overlayView.setResults(handLandMarkerResult, bitmap.height, bitmap.width)
-//                overlayView.draw(canvas)
-                utils.showToast("반환")
+                val resizedImg = Bitmap.createScaledBitmap(image, ivImage.width, ivImage.height, true)
+                val canvas = Canvas(resizedImg)
+                initPaints()
+                setResults(handLandMarkerResult, resizedImg.height, resizedImg.width)
+                draw(canvas)
+                ivImage.setImageBitmap(resizedImg)
+
+                utils.showToast("결과 사진 업로드 완료.")
             } else
                 utils.showToast("분석을 수행할 수 없습니다.")
+        }
+    }
+
+    // ***** 추출한 LandMark를 이미지에 그리는 메소드 ******
+    // 1. initPaints() : 초기값 설정 메소드
+    private fun initPaints() {
+        linePaint.color =
+            ContextCompat.getColor(this, R.color.mp_color_primary)
+        linePaint.strokeWidth = LANDMARK_STROKE_WIDTH
+        linePaint.style = Paint.Style.STROKE
+
+        pointPaint.color = Color.YELLOW
+        pointPaint.strokeWidth = LANDMARK_STROKE_WIDTH
+        pointPaint.style = Paint.Style.FILL
+    }
+
+    // 2. setResults() : 변수 값 설정하는 메소드
+    private fun setResults(
+        handLandmarkerResults: HandLandmarkerResult,
+        imageHeight: Int,
+        imageWidth: Int,
+        runningMode: RunningMode = RunningMode.IMAGE
+    ) {
+        results = handLandmarkerResults
+
+        this.imageHeight = imageHeight
+        this.imageWidth = imageWidth
+
+        scaleFactor = when (runningMode) {
+            RunningMode.IMAGE,
+            RunningMode.VIDEO -> {
+                min(ivImage.width * 1f / imageWidth, ivImage.height * 1f / imageHeight)
+            }
+
+            RunningMode.LIVE_STREAM -> {
+                // PreviewView is in FILL_START mode. So we need to scale up the
+                // landmarks to match with the size that the captured images will be
+                // displayed.
+                max(ivImage.width * 1f / imageWidth, ivImage.height * 1f / imageHeight)
+            }
+        }
+    }
+
+    // 3. draw() : 매개변수로 주어진 canvas에 실제 LandMark를 그리는 메소드
+    private fun draw(canvas: Canvas) {
+        results?.let { handLandmarkerResult ->
+            for (landmark in handLandmarkerResult.landmarks()) {
+                // 1. 추출된 LandMark를 점으로 표시
+                for (normalizedLandmark in landmark) {
+                    canvas.drawPoint(
+                        normalizedLandmark.x() * imageWidth * scaleFactor,
+                        normalizedLandmark.y() * imageHeight * scaleFactor,
+                        pointPaint
+                    )
+                }
+                // 2. LandMark들끼리 선으로 연결
+                HandLandmarker.HAND_CONNECTIONS.forEach {
+                    canvas.drawLine(
+                        landmark[it!!.start()]
+                            .x() * imageWidth * scaleFactor,
+                        landmark[it.start()]
+                            .y() * imageHeight * scaleFactor,
+                        landmark[it.end()]
+                            .x() * imageWidth * scaleFactor,
+                        landmark[it.end()]
+                            .y() * imageHeight * scaleFactor,
+                        linePaint
+                    )
+                }
+            }
         }
     }
 }
