--- app/src/main/java/com/example/borntobe/OverlayView.kt
@@ -1,111 +1,0 @@
-package com.example.borntobe
-
-import android.content.Context
-import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.Paint
-import android.util.AttributeSet
-import android.view.View
-import androidx.core.content.ContextCompat
-import com.google.mediapipe.tasks.vision.core.RunningMode
-import com.google.mediapipe.tasks.vision.handlandmarker.HandLandmarker
-import com.google.mediapipe.tasks.vision.handlandmarker.HandLandmarkerResult
-import kotlin.math.max
-import kotlin.math.min
-
-class OverlayView(context: Context?, attrs: AttributeSet?) :
-    View(context, attrs) {
-    companion object {
-        private const val LANDMARK_STROKE_WIDTH = 8F
-    }
-    private var results: HandLandmarkerResult? = null
-    private var linePaint = Paint()
-    private var pointPaint = Paint()
-    private var scaleFactor: Float = 1f
-    private var imageWidth: Int = 1
-    private var imageHeight: Int = 1
-
-    // init : 생성자를 통해 인스턴스가 만들어질 때 호출
-    init {
-        initPaints()
-    }
-
-    // clear() : 기존 변수 및 메소드 초기화 하는 메소드
-    fun clear() {
-        results = null
-        linePaint.reset()
-        pointPaint.reset()
-        invalidate()
-        initPaints()
-    }
-    
-    // initPaints() : 초기값 설정 메소드
-    private fun initPaints() {
-        linePaint.color =
-            ContextCompat.getColor(context!!, R.color.mp_color_primary)
-        linePaint.strokeWidth = LANDMARK_STROKE_WIDTH
-        linePaint.style = Paint.Style.STROKE
-
-        pointPaint.color = Color.YELLOW
-        pointPaint.strokeWidth = LANDMARK_STROKE_WIDTH
-        pointPaint.style = Paint.Style.FILL
-    }
-
-    // draw() : 매개변수로 주어진 canvas에 실제 LandMark를 그리는 메소드
-    override fun draw(canvas: Canvas) {
-        super.draw(canvas)
-        results?.let { handLandmarkerResult ->
-            for (landmark in handLandmarkerResult.landmarks()) {
-                // 1. 추출된 LandMark를 점으로 표시
-                for (normalizedLandmark in landmark) {
-                    canvas.drawPoint(
-                        normalizedLandmark.x() * imageWidth * scaleFactor,
-                        normalizedLandmark.y() * imageHeight * scaleFactor,
-                        pointPaint
-                    )
-                }
-                // 2. LandMark들끼리 선으로 연결
-                HandLandmarker.HAND_CONNECTIONS.forEach {
-                    canvas.drawLine(
-                        landmark[it!!.start()]
-                            .x() * imageWidth * scaleFactor,
-                        landmark[it.start()]
-                            .y() * imageHeight * scaleFactor,
-                        landmark[it.end()]
-                            .x() * imageWidth * scaleFactor,
-                        landmark[it.end()]
-                            .y() * imageHeight * scaleFactor,
-                        linePaint
-                    )
-                }
-            }
-        }
-    }
-
-    // setResults() : 변수 값 설정하는 메소드
-    fun setResults(
-        handLandmarkerResults: HandLandmarkerResult,
-        imageHeight: Int,
-        imageWidth: Int,
-        runningMode: RunningMode = RunningMode.IMAGE
-    ) {
-        results = handLandmarkerResults
-
-        this.imageHeight = imageHeight
-        this.imageWidth = imageWidth
-
-        scaleFactor = when (runningMode) {
-            RunningMode.IMAGE,
-            RunningMode.VIDEO -> {
-                min(width * 1f / imageWidth, height * 1f / imageHeight)
-            }
-            RunningMode.LIVE_STREAM -> {
-                // PreviewView is in FILL_START mode. So we need to scale up the
-                // landmarks to match with the size that the captured images will be
-                // displayed.
-                max(width * 1f / imageWidth, height * 1f / imageHeight)
-            }
-        }
-        invalidate()
-    }
-}
